pyproject.toml
[project]
name = "ninja-keys"
version = "1.0.0"
description = "API Keys for Django Ninja"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Felipe R. de Almeida", email = "felipe@security.dev" }
]
requires-python = ">=3.9"
classifiers = ["Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Web Environment",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1", ]
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "django-ninja>=1.3.0",
    "django-test-migrations>=1.4.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.10.0",
    "ruff>=0.9.6",
]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
src = ["src", "test_project", "tests"]
